Remove VCS providers
We tried to support multiple VCS providers and due to how some written this does not give the best user experience. We tried to several work arounds but ultimately VCS repos should be managed outside of this module and passed into this module creating an implicit dependency so they're created before the workspace is created (and then attached to).

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>