Allow TFC agents to use an IAM role to manage account resources
Currently we create an IAM user per workspace in each account to manage
AWS resources. This method creates access keys and secret keys in the
managed account and adds overhead of rotating these long lived keys and
the risk of someone getting access to the keys.

By instead using a role we no longer need to manage these keys or be at
risk of them being compromised, and can restrict the role so that it can
only be assumed by the account and role used by the TFC agents. We also
create a random external ID per workspace role so that should someone
get access to the TFC agent role they cannot easily assume roles in
managed accounts.

Co-authored-by: Stephen Hoekstra <shoekstra@schubergphilis.com>